requirements.txt
playwright
pytest
pytest-html
allure-pytest
pandas
openpyxl
